+++ app/Harvest/Importers/ItemImporter.php
@@ -42,8 +42,13 @@ class ItemImporter extends AbstractImporter
         return $this->mapper->mapId($row);
     }
 
-    protected function processBelongsToMany(Model $model, $field, array $relatedRows, $createRelated = true) {
-        $createRelated &= !in_array($field, ['authorities', 'collections']);
-        parent::processBelongsToMany($model, $field, $relatedRows, $createRelated);
+    protected function processHasMany(Model $model, $field, array $relatedRows, $allowDelete = true) {
+        $allowDelete &= !in_array($field, ['images']);
+        parent::processHasMany($model, $field, $relatedRows, $allowDelete);
+    }
+
+    protected function processBelongsToMany(Model $model, $field, array $relatedRows, $allowCreate = true) {
+        $allowCreate &= !in_array($field, ['authorities', 'collections']);
+        parent::processBelongsToMany($model, $field, $relatedRows, $allowCreate);
     }
 }